server/requirements.txt
aniso8601==9.0.1
cbor2==5.4.3
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Flask==2.1.3
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
luma.core==2.3.1
luma.led-matrix==1.6.1
MarkupSafe==2.1.1
Pillow==9.2.0
pkg-resources==0.0.0
pyftdi==0.54.0
pyserial==3.5
python-dotenv==0.20.0
pytz==2022.1
pyusb==1.2.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.2
spidev==3.5
typing-extensions==4.3.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
